+++ apps/admin/src/features/heroImagesManager/hooks/useCreateHeroImage.ts
@@ -1,6 +1,6 @@
 import { useCarouselServicePostApiV1Carousels } from '~/apis/admin/queries'
 import { useCarouselServiceGetApiV1CarouselsKey } from '~/apis/community/queries'
-import { queryClient } from '~/utils/get-query-client'
+import { queryClient } from '~/shared/utils/'
 import { MESSAGES } from '../constant/constants'
 
 import type { MessageInstance } from 'antd/es/message/interface'
