+++ apps/admin/src/main.tsx
@@ -6,7 +6,7 @@ import '@ant-design/v5-patch-for-react-19'
 
 import { QueryClientProvider } from '@tanstack/react-query'
 import { routeTree } from './routeTree.gen'
-import { queryClient } from './utils/get-query-client'
+import { queryClient } from './shared/utils'
 
 import './apis/admin/OpenAPIConfig'
 import './styles/globals.css'
