+++ apps/admin/src/routes/news/index.tsx
@@ -1,8 +1,7 @@
 import { createFileRoute, useSearch } from '@tanstack/react-router'
 import { Suspense } from 'react'
 
-import { SearchBar } from '~/components/user/search-bar'
-import { PostList } from '~/shared/components/Post'
+import { PostList, SearchBar } from '~/shared/components/Post'
 
 import { PATH } from '~/constants/path'
 import { useNewsList } from '~/shared/hooks'
