+++ apps/admin/src/routes/notice/new/index.tsx
@@ -1,10 +1,10 @@
 import { createFileRoute } from '@tanstack/react-router'
-import { WriteNewPostField } from '~/components/posts/write-new-post-field'
+import { PostForm } from '~/shared/components/Post'
 
 export const Route = createFileRoute('/notice/new/')({
   component: NoticePostPage,
 })
 
 function NoticePostPage() {
-  return <WriteNewPostField />
+  return <PostForm />
 }
