+++ apps/admin/src/routes/professor/index.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router'
-import { ProfessorCreator } from '~/components/professor/professor-creator'
-import ProfessorTable from '~/components/professor/professor-table'
+import { ProfessorCreator } from '~/features/professor/components/ProfessorCreator'
+import ProfessorTable from '~/features/professor/components/ProfessorTable'
 
 export const Route = createFileRoute('/professor/')({
   component: ProfessorPage,
