+++ apps/admin/src/shared/components/Post/Board.tsx
@@ -14,7 +14,7 @@ import { EDIT_POST_PATH_MAP, type PostCategory } from '~/constants/path'
 import { useModal } from '~/hooks/use-modal'
 import { MESSAGES } from '~/shared/constant/post.constants'
 import { useDeletePost } from '~/shared/hooks/Post'
-import { extractFileName } from '~/utils/utils'
+import { extractFileName } from '~/shared/utils/utils'
 
 interface HeaderProps {
   title: string
