+++ apps/community/src/app/(about)/club/page.tsx
@@ -3,7 +3,7 @@ import { HydrationBoundary, dehydrate } from '@tanstack/react-query'
 import { ClubList } from '~/features/club/components/club-list'
 import { CLUB_QUERY_OPTIONS } from '~/features/club/services/queries'
 import { PageHeader } from '~/shared/components/page-header/page-header'
-import { getQueryClient } from '~/shared/utils/get-query-client'
+import { getQueryClient } from '~/shared/utils/'
 
 export const dynamic = 'force-dynamic'
