+++ apps/community/src/app/(board)/news/page.tsx
@@ -4,7 +4,7 @@ import * as styles from '~/app/(board)/notice/page.css'
 import { SearchBar } from '~/features/board/components/search-bar'
 import { BOARD_QUERY_OPTIONS } from '~/features/board/services/queries'
 import { PageHeader } from '~/shared/components/page-header/page-header'
-import { getQueryClient } from '~/shared/utils/get-query-client'
+import { getQueryClient } from '~/shared/utils/'
 import { PaginatedBoardList } from '~/widgets/board/components/paginated-board-list'
 
 export const dynamic = 'force-dynamic'
