+++ apps/community/src/app/lab/page.tsx
@@ -3,7 +3,7 @@ import { HydrationBoundary, dehydrate } from '@tanstack/react-query'
 import { LabList } from '~/features/lab/components/lab-list'
 import { LABS_QUERY_OPTIONS } from '~/features/lab/services/queries'
 import { PageHeader } from '~/shared/components/page-header/page-header'
-import { getQueryClient } from '~/shared/utils/get-query-client'
+import { getQueryClient } from '~/shared/utils/'
 
 export const dynamic = 'force-dynamic'
