+++ apps/community/src/app/professor/page.tsx
@@ -3,7 +3,7 @@ import { HydrationBoundary, dehydrate } from '@tanstack/react-query'
 import { ProfessorList } from '~/features/member/components/professor-list'
 import { PROFESSORS_QUERY_OPTIONS } from '~/features/member/services/queries'
 import { PageHeader } from '~/shared/components/page-header/page-header'
-import { getQueryClient } from '~/shared/utils/get-query-client'
+import { getQueryClient } from '~/shared/utils/'
 
 export const dynamic = 'force-dynamic'
